Add a graceful_exit option to Worker
Worker handles signals by completing the current job and then exiting.
For our purpose, we need the Worker to complete *all* outstanding
jobs before terminating. In order to enable this behaviour, set the
`graceful_exit` option to `true`, or set the `GRACEFUL_EXIT`
environment option if using the Rake task.